refactor(editor-toolbar): Split monolithic toolbar into modular components
- Extract DocumentMenu component for file-level actions
- Extract FormattingToolbarGroup component for text formatting and table operations
- Extract OverflowMenu component for Vim/theme controls and settings
- Extract PipelinesToolbarGroup component for draggable transformation pipelines
- Extract ToolbarButton component for reusable button with tooltip
- Move EditorToolbarProps interface to dedicated types file
- Simplify main EditorToolbar component to compose subcomponents
